A few tweaks to Thomas's solution
Thomas's solution is correct and works well, but I prefer to work
with the original point objects (with their member methods, such as
they are, intact) rather than with a 3-tuple of (x,y,on_curve). So
I've tweaked his algorithm to use Fontforge point objects. (And to
use a generator function to yield pairs of points to examine.)

I've also restored the original extractglyph.py program that we
wrote to extract glyphs to SVG format, because looking at those
glyphs in Inkscape has proven handy for debugging.