Extract one outline from a font, turn it into vectors, then into an svg polyline
Note that we're "cheating" in several places: we just draw vectors
between the on-curve control points of the Bezier curve without checking
the shallowness of the resulting angles. If a font had (say) a circle
with very few control points, this would turn it into a diamond.

We also only extract the first outline from the glyph: in the next
version, we'll extract all outlines.

But hey... it works!