Added demo of triangulation
This relies on the python-poly2tri demo being installed alongside
the current repo (clone https://github.com/hansent/python-poly2tri
as a sister folder to the one in which this repo was cloned). For
now, we're just handing the triangulation code off to poly2tri.
Eventually we will do the triangulation ourselves, *properly*, and
even draw the results ourselves (using pygame to draw the results
just as poly2tri does). But for now, doing it this way lets us look
at the results and see what the triangulation looks like, and get a
feel for what drawing the midpoints of those triangles will do.