WIP: Figuring out why endpoints aren't found
It seems the approach of turning triangles into Decimals and putting
them into sets is failing for some cases: by plotting fat blue points
at all identified midpoints, we can see that a couple of them are on
edges of the glyph, which should have been thrown out by the algorithm.
Since that isn't working, we'll have to scrap the approach of using sets
and instead use an O(N) search through the identified edge vectors.