WIP: Trying an angle-based approach
We'll look at the "second derivative" of midlines (calculated by
looking at the angles between the last few segments of the lines)
and continue the line by the following method:

1) First, extend the previous several lines *tentatively* with the
angle being the average of the previous N angles (so if the progression
is -3, -5, -7, then the next few lines will be at -5, -5, -5). Each line
segment will have the same length as the average length of the previous
few line segments, so that the rough shape is a circular arc. This will
extend by several line segments.

2) Look at all the "candidate" segments, and see where they would
intersect with other "candidates" or with real segments. Trace them
until those intersection points. Delete the rest of the candidates,
the part that would extend past that intersection.

3) Turn all the candidates into real segments, and draw them.