WIP: Visual of angle-based approach
This draws an extra-long line (500 pixels long) so that it's easy
to see where the intersection points will be. The angle of this line
to the previous one is calculated by taking the average of the previous
5 angles (skipping the very last one, since I found that the last line
segment was usually distorted a bit compared to the other ones).

However, when running this we'll see that the approach is not working
very well. I may pursue this approach a bit further, but I think this
experimental branch is going to be a dead end, unfortunately.