Code cleanup: Remove unused functions
While iterfunc, itermap, iterfiltermap and the like were very useful
exercises for a beginning Python programmer, we're not actually using
them in the rest of our code. They closely duplicate the functionality
of the standard library's itertools.imap and itertools.ifilter, and/or
the functionality available in list comprehensions, and when we need
that functionality we're using the standard library functions. So this
code is taking up space for no current benefit.