Remove another unused function
closestpoint_in_same_direction was part of an algorithm we've since
abandoned: it traced lines between midpoints by looking for the closest
point in the "same direction" as the incomplete line already traced,
where "same direction" was defined as "the angle shouldn't vary by more
than N degrees" (N defaulting to 30). That algorithm has since been
abandoned, and we're not going to go back to it -- the "find triangles
sharing a common side" algorithm works much better. So we can delete all
the code relating to the "closest point in same direction" algorithm.