Remove most commented-out code
There are a couple of abandoned approaches still lingering in
commented-out sections of code, but this is making it a bit harder to
identify functions that are truly unneeded. Deleting all the
commented-out code will make "orphaned" (unneeded) functions a little
more obvious, as they can now be identified by a simple text search.