Remove two more unused functions
get_glyph() was from an early version of the command-line interface, now
superseded. find_straight_lines() was from an early approach to finding
the glyph's "stroke width" (that is, the usual distance between parallel
sides -- or "parallel" curves -- of the glyph's outline). The approach
we're taking now uses Shapely to calculate the stroke width, and seems
much more reliable, so we can get rid of find_straight_lines() as well.