Silence FontForge's complaints about opened fonts
When opening a font with fontforge.open(), FontForge gets very verbose
about issues found in the font. This is a good thing when the font is
being opened in the FontForge editor, because it gives the user the
opportunity to correct those errors. But since we're just trying to
process the font automatically in Python, we don't want to see all those
error messages, so we'll suppress stderr while opening font files.