Proof-of-concept: output one dotted glyph to TTF
This is the first step towards creating a real TTF containing the dotted
font. For this proof-of-concept:
  * The output file name cannot be changed; it is always "output.ttf"
  * Only a single glyph will be created in the output TTF file
  * The output TTF has no font name, hinting, or anything but one glyph

In the final version, the output name can be selected but will default
to "originalfilename_dotted.ttf", and the output TTF will have most
relevant data copied from the original font file. (Some things, like
hints, will not be copied because they are based on the original
outlines, which are completely different in the dotted TTF. And since
the dotted TTF is intended for printing, hints are pointless anyway.)