Change cmdline parameters to prepare for --output
We will soon want to use -o / --output to specify the output filename.
To free up the -o option, we'll change --hide-outline to --show-glyph.
It needed to change from "hide" to "show" anyway, to provide consistency
with the rest of the options, and this is a good time to make the change.