New approach to create new font: copy, then change
Instead of creating a new font from scratch, we'll copy the TTF file
provided and use that as a basis for the font creation process. This
will mean that all sorts of things will be copied over, like font name,
copyright information, etc., which may need to be modified. But it's the
best way of ensuring that glyph anchor points, lookup subtables, and so
on are copied into the new font.