Add subprocess32 module (third-party)
This module comes from https://pypi.python.org/pypi/subprocess32/. It's
a backport of the Python 3.2 version of the subprocess module to 2.7;
the main reason for using this backport is for the timeout parameter to
subprocess.Popen.communicate(). Without the timeout, if the child
process gets stuck in an infinite loop, we don't have a good way to
detect this.

We'll use this module in the next commit, to implement a safer way to
call cdt.triangulate() that won't get stuck if it goes into an infinite
loop.