Make triangles in a separate process
Because some badly-designed glyphs can send the cdt.triangulate() call
into an infinite loop, we'll use a slower (but safer) method to call
cdt.triangulate(): we'll call it in a separate process, and kill that
process if it takes too long.

Eventually, we should turn this into a command-line option so the user
has the choice of whether to use "safe mode" or "speedy mode".